Increase the limit for subreddit post fetching
Increased the number of fetched posts from Reddit's subreddit from 1 to 5. This change will allow the bot to provide users with more diverse content. As part of this change, also ensured that new submissions are handled within the comment stream. This modification improves real-time post handling and subsequent user interactions.