Ignore additional cache files in repo
The commit updates the `.gitignore` file to include additional /cache files. These files were added to the ignore list because they are backup files and not necessary for the proper functioning of the software, thereby preventing unnecessary bloating of the repo. This change prevents these files from being tracked and included in future commits.