Add new scripts for Reddit bot and improve image captioning
In this commit, two newPython scripts are added for the Reddit bot. The 'spot_post.py' script enables the bot to post images on the subreddit 'wallstreetbets' with flair text. 'training_clean_up.py' improves the bot's image captioning capability through the use of BLIP (short for BERT and Transformers on Images and Text) for conditional image captioning.

Unused modules from a previous implementation are removed. The sleep time in the 'run-test-image-generation.py' is adjusted for optimization purposes. Some file pathnames are also added to the '.gitignore' file.

These changes enhance the Reddit bot's functionality in terms of processing Reddit threads and images for improved user interaction while optimizing its general performance.