Remove redundant file and refactor gather_sync.py
The script reddit-collect-comment-data-sync.py is removed as it's redundant and being replaced. The gather_sync.py is accordingly updated to refine the structure of classes, reposition Enum class SubmissionType and fine-tune the handling of Reddit submissions. The model location is updated and mining from multiple subreddits is made possible for better versatility. These changes aim to streamline the data gathering process and provide clearer, more maintainable code structure.